Remove use of Model.scoped [#174954960]
In Rails 4 Model.scoped is deprecated in favor of using Model.all which now returns a relation and thus can be chained.  We were using Model.scoped in the Pundit policy scope to convert all to a relation.

This also updates the none method to use the new Rails 4 Model.none method instead of the empty query hack needed in Rails 3.