175078001 - update Report::EmbeddableFilter
Report::EmbeddableFilter serializes fields.
It seems that `#read_attribute()` now returns nil for fields that are not set.  The implementation and test seem to indicate that unset values were previously stored as an itterable.

This update adds a nil check, and exits early if the read-cache is present.

As per the comment
>  newer version of mock_model take care of this for you

[#175078001]
https://www.pivotaltracker.com/story/show/175078001