allow resource_link_id param when requesting JWT for a teacher
This fixes an issue in the portal-report. The report doesn't know
if it is being launched as a teacher or a student. So it will request a JWT
with a resource_link_id regardless of what type of user.

Since the JWT controller has a resouce_link_id for a teacher it seemed best
to validate that this teacher had access to it. And also use it to add the class_hash
to the JWT even if the class_hash param was not provided.

[#177080369]