Fixed save of enabled bookmark types in admin settings [#177438793]
The strong  params setting for enabled_bookmark_types was not set as an array value so it was being dropped before it reached the update_attributes method in the settings model which used it.

I checked the other models and this is the only field that is serialized as an array so no other strong params need to be updated.