Fixed  class sort api url [#177438813]
The class sort api url was set as a member resource in the routes which was not correct (as it pertains to all classes and not a specific class).

This caused the "Manage Class List" to silently fail which wasn't evident until you reloaded the page and saw the class orderings revert.