separate out app:setup:create_default_data
This had been broken since I cleaned up db:features_test:prepare
The app:setup:create_default_data task was running in the development environment.
It was populating the cucumber database because of the active record statements in db:features_test:prepare
But the Rails.env was development. This meant the password hashing algorithm was
the same as development so it had 10 stretches. But at the same time the spec_helper was loaded
which meant that pepper was being overridden.

The result of those password issues was that the users created by the rake task had invalid
passwords when the cucumber tests were actually running. We had just worked around this
by cutting out the manual login of those users without actually fixing the problem.