Add tests for jwt routes
This adds the authorization tests for various users hitting the two new
`report_learners_es_controller` routes, and confirms that an admin user
with a valid JWT can receive an array of detailed learners.

It doesn't cover the specifics of the permissions-form access, since
that's covered in slightly more depth in the original `external_report_query`
tests, which uses the same ES function.