Include `query` in `get` requests. Empty hash params are `nil`
There was a change in how controller tests are written.  You now have to specify the `:params` explicitly.

Also when a parameter is sent as an empty hash from the test, the controller thinks its nil now.
There is a github issue for this: https://github.com/rails/rails/issues/26569

#178382747 - merge upstream changes into Rails6
https://www.pivotaltracker.com/story/show/178382747