Split teachers for report into array if multiple
If a class has multiple teachers, the `teachers_` properties will be
a comma-separated list. This splits the teachers into an array of
multiple teachers if so.

I tried breaking the splitting by adding a comma to the school district
and to one of the teacher's names, and it handled it fine, substituting
the commas with spaces. (From Kibana I can see that the values are
being saved in ES without the commas.)