Update postToUrl function to accept five arguments
Previously, in 1adfab777, I updated the call to `postToUrl` to pass in
five arguments, which included a new jwt created in the researcher
request. However, I did not update the `postToUrl` to accept the new
arguments. This went unnoticed because the Learners report only used the
first four arguments.

This exposes a redundancy in the code, which is that the researchers
report is already generating a JWT which is functionally identical to
the one passed in by the ReportLearnersEsController. This is probably
not necessary and could be simplified.

(Simplification would involve updating all the new reports, though,
which currently expect this `jwt` parameter.)