Re add username and student_name to learners
These two fields were lost when the code was converted to use the ES document instead of the report learner
Rather than add the fields to the ES document, they are pulled from the portal database this way
Elastic search doesn't contain the most sensitive student names.

The skip_report_learners option was replaced with a learner_type option which seemed more clear.
We are planning to completely remove the Report::Learner model in the future.

[#178897764]