optimize the query to update Elasticsearch
This query was causing an out of memory errors.
It would load in 1000 offerings and then include all of their learners and related objects.

Now it starts with 100 offerings and does not initially include the learners from the offering.
Then the learners of the offering are loaded in batches default for find_each is 1000 learners.

The includes are split up, to reduce redundant queries.