Add codebuild config buildspec.yml
This also attempts to setup the Dockerfile to pickup the version environment from a build arg
This way the version info is embedded in the docker image and isn't needed to be passed in.
I'm pretty sure the version support will require additional changes so this doesn't get overridden
by a blank environment setting.