Only open material edit form in an iframe for teacher users.
PT Story: https://www.pivotaltracker.com/story/show/179197451

[#179197451]

The requirements listed in the story above are already met except for having the material edit form open in an iframe only for non-author users. These changes ensure that the iframed edit form is only shown to non-author and non-admin users. There is also a small style tweak here that allows the iframe to take up the full available width of the page. That makes it easier for teacher authors to see and access the "edit" links in the ITSI-style edit form.