fix error during error reporting, and log exception
For some reason we are getting students that are not being able log in.
Our exception notifier is providing an error with:

```
A NameError occurred in authentications#google:

  undefined local variable or method `env' for #<AuthenticationsController:0x00000000d87200>
Did you mean?  end
               END
  app/controllers/authentications_controller.rb:81:in `rescue in generic_oauth'
```

Which I think means the env needs to be changed to `request.env`.
However fixing this also means that we wouldn't see these exceptions anymore.
So I added the ExceptionNotifier call to record the actual exception.

The existing recorded exceptions include email addresses that don't exist in the database.

I suspect it could be one of 2 things:
- the authentication with google is taking a while, so the student gets impatient and reloads the page, this causes some error.
- the student names or emails have some kind of invalid characters in them and this causes an exception while trying to create
the user object