Add provider and uid
the recent error messages are reporting non-unique email and login fields
when creating the User. These fields are based on the provider and uid.
Including them here will let us track down the existing user to figure out how it
is broken.

In theory any user with an email and login that matches should have an
Authentication object too.
If that Authentication exists then User#find_for_omniauth should find it
first and return the associated user.
But for some reason this isn't working.