add indexes to authentications and access_grants
The SSO code looks for authentications based on the provider and uid.
The production site has 428,000 authentications, so this is a slow process

When LARA is using the portal for SSO, the portal has to look up the
access grant by the access token. This was not indexed and there are
56,000 access grants on production, so this is slow.