Add sitewide alert banner on learn.concord.org
PT Story: https://www.pivotaltracker.com/story/show/182613149

[#182613149]

Adds a new field to admin/portal settings for sitewide alert banner content. When content is present, a banner appears at the top of every page containing that content. The banner is dismissable. If a user dismisses the banner, a cookie is set so we can know not to show it to them again.

To do: If we decide to keep this feature as is, we will need to modify it to provide a unique cookie name to go with every alert's content so even if a user dismisses one alert, they will see future alerts about other issues.