Added user and portal_url to external report queries [#180951212]
Previously the info of the user making the request was only available in the learner query.  This makes the user info available in both the learner and user external report queries as it is needed to create the AWS workgroup.

This also adds a new portal_url attribute to both the learners and user queries.  The current query creator code infers this url from another url attribute (learnersApiUrl) and since it is also needed in the user report handling it has been pulled out into its own attribute.