feat: Add supports_researchers flag to external reports [PT-187221928]
The supports_researchers notes if the external report supports anonymized researcher views.

This flag is returned in the offerings api result.

This commit adds the flag to the mapped external reports data available to the React components but does not add logic to filter out the reports.

The new researcher views React components will need to filter out reports in future commits.