fix: Users always redirected to learner report path on report server [PT-166155743]
This happens both on login after clicking on a report from the main portal and when using the report portal as an oauth client.

The change here is to move the existing redirect to the learner report to the user's current homepage helper on the report server and removing the check for the report server in the application_controller.rb#after_sign_in_path_for method.