fix: Collection lightbox header not showing current collection [PT-187727240]
This fixes a bug introduced in the React 18 upgrade.  The lightbox nav was holding the selected collection in its state, initially set via props.  This broke in the React 18 upgrade due to how the lightbox is now shown and it was not needed so the copy into state was removed.

This also fixes two unrelated semicolon lint errors that were clogging up the PR view.