fix: Allow project admins of any project to toggle hide names control [PT-187796954]
Previously only site admins and managers could toggle the hide names control in the learner reports.  This expands this ability for project admins of any project to also toggle this control.

The original reason that only site admins and managers had access is the user could use the learner report page to select teachers or resources both in and out of a project they admined so there was not enough fine grained control.  It was decided to open this toggle up to any project admin as those roles have blanket coverage in the IRBs.