fix: Passwords validations in signup form [PT-187798321] [PT-187798313]
With the update to React 18 the Formsy input validations broke as they relied on state updates being done before the validations.

This commit changes the password confirmation validation to use a function that compares the passed form values.  It also updates the the password length validations to remove the max length validation.