fix: Show archived classes in permissions UI [PT-188564811]
This removes the server-side filtering of archived classes and adds the is_archived attribute to the returned class list.  A new column called "Class Status" has been added to the class list when a teacher in selected in the permissions UI.  This new column shows "Archived" if the class is archived and "Active" is it is not.