fix: Revert default teacher classes list to not include archived classes [PT-188564811]
In case other existing api consumers are expecting non-archived classes to not be in the list of the teacher's classes a new query parameter, include_archived, must be set to true to include the archived classes in the list.

The permission forms component has been updated to use this new parameter.