feat: Add researcher portal JWTs [PT-188754102]
This adds a new user_type of "researcher" to the portal JWTs.  Like the Firebase JWT code the requester needs to add "researcher=true" to the request to ask for a researcher JWT.

Unlike the Firebase JWT no checks are made to see if the user is actually a researcher.  This is because, unlike the Firebase JWT, no claim is being made for a specific class and the researcher user_type is used only by clients to know what type of JWT they have and all requests to the portal using the JWT are checked for permissions using the user_id which will check if the user has permissions to access the data.

These JWTs are used currently only by generated links in reports and those links may be used by people who are not researchers but have access to the data as project admins.