feat: Add teacher feedback notice to student offerings [PT-188732128]
This adds a new api endpoint that calls a Firebase function using an url provided in the environment variables, along with a source key and bearer authentication token.  That Firebase function returns a map of offering ids to feedback metadata, that for now, includes the last time feedback was created/updated for the offering by a teacher.