feat: Add sign in or register page [RIGSE-145]
This adds a new page that shows a login button and a create account button.

The page accepts a login_url query parameter that is used instead of calling the portal api.  CLUE uses this for standalone mode login to redirect back to itself so it can start the oauth flow.

The page also accepts a class_word parameter that is used to pre-fill the class word field in the signup form.  This is used by CLUE to pass the class word in the standalone url to the signup form.

The page is accessible at /sign-in-or-register and is not linked anywhere in the portal.  It was designed however to replace the two login and register buttons in the top right corner of the portal.

Changes in this commit:

- Added a new page at /sign-in-or-register that shows a login button and a create account button.
- Updated the routes to include the new page.
- Updated the react components to include the new page.
- Updated the react components to pass the login_url and class_word to all the needed components.
- Updated the login buttons to first check for the login_url before calling the portal api.