fix: Url params getting unescaped during signup
In the signup process when the user picked if they were a student or teacher the code was updating the userType query param in a way that caused the other params to get unescaped.

The updated code uses the modern URL class to set the userType instead of doing it with building the url manually.