feat: Add per student setting of active and locked offerings [RIGSE-207]
This adds a new model, UserOfferingMetadata, that stores the active and locked flags per user+offering.

A new model is needed as the learner model is only created when a student runs an activity and these flags are checked before a student runs an activity.

This commit also fixes some existing lint errors to clean up the test output.