feat: Update footers [RIGSE-238]
This updates the copyright notice to include "The" before "Concord Consortium", changes "All rights reserved" to "All Rights Reserved", and adds a note about the 501(c)(3) status.

This also adds the React version to the NGSS footer and simplifies the current copyright year display of both the default and NGSS footers.