allow project admins to view the roster
I also allowed project admins to view the researcher view of the materials page.
We plan to give project admins access to the "Research Projects" menu, so they can access the roster from there. Since we are doing that it seemed worthwhile to let them also access the materials page.