feat: Added Teacher Guide feature [RIGSE-300]
This commit introduces a new Teacher Guide page. It includes the ability to create and preview a custom teacher guide page, as well as an option to link to an external URL for the teacher guide.

The necessary model, controller, views, and migration files have been added and modified accordingly.

The settings page has also been updated to use fieldsets and legends for better organization of content sections.